refactor(materializer): extract graph-lifecycle status/lifecycle mappers (file #3)
Lift mapGraphStatus, materializeLifecycle, buildLifecycleActions,
displaySafeDegradationReason into graph-lifecycle.ts. Pure functions of the
canonical source. Net 74/74 green, check 0.
