+++ packages/desktop/src/lib/acp/session-state/graph-lifecycle.ts
@@ -1,0 +1,117 @@
+/**
+ * Lifecycle/status projection helpers for the canonical agent-panel graph
+ * materializer: canonical session status, lifecycle model, lifecycle actions,
+ * and degradation-reason display text. Pure functions of the canonical source —
+ * no mutation, no side effects. GOD-safe.
+ */
+import {
+	AGENT_PANEL_ACTION_IDS,
+	type AgentPanelActionDescriptor,
+	type AgentPanelLifecycleModel,
+	type AgentPanelSessionStatus,
+} from "@acepe/ui/agent-panel/types";
+import type { OperationDegradationReason } from "../../services/acp-types.js";
+import type { AgentPanelCanonicalSource } from "./agent-panel-canonical-source.js";
+
+export function mapGraphStatus(graph: AgentPanelCanonicalSource): AgentPanelSessionStatus {
+	const lifecycleStatus = graph.lifecycle.status;
+	if (
+		lifecycleStatus === "failed" ||
+		graph.activity.kind === "error" ||
+		graph.turnState === "Failed"
+	) {
+		return "error";
+	}
+	if (
+		lifecycleStatus === "reserved" ||
+		lifecycleStatus === "activating" ||
+		lifecycleStatus === "reconnecting"
+	) {
+		return "warming";
+	}
+	if (lifecycleStatus === "detached" || lifecycleStatus === "archived") {
+		return "idle";
+	}
+	if (
+		graph.activity.kind === "running_operation" ||
+		graph.activity.kind === "awaiting_model" ||
+		graph.turnState === "Running"
+	) {
+		return "running";
+	}
+	if (graph.turnState === "Completed") {
+		return "done";
+	}
+	return graph.transcriptSnapshot.entries.length > 0 ? "idle" : "connected";
+}
+
+export function materializeLifecycle(graph: AgentPanelCanonicalSource): AgentPanelLifecycleModel {
+	return {
+		status: graph.lifecycle.status,
+		detachedReason: graph.lifecycle.detachedReason ?? null,
+		failureReason: graph.lifecycle.failureReason ?? null,
+		errorMessage: graph.lifecycle.errorMessage ?? null,
+		actionability: {
+			canSend: graph.lifecycle.actionability.canSend,
+			canResume: graph.lifecycle.actionability.canResume,
+			canRetry: graph.lifecycle.actionability.canRetry,
+			canArchive: graph.lifecycle.actionability.canArchive,
+			canConfigure: graph.lifecycle.actionability.canConfigure,
+			recommendedAction: graph.lifecycle.actionability.recommendedAction,
+			recoveryPhase: graph.lifecycle.actionability.recoveryPhase,
+			compactStatus: graph.lifecycle.actionability.compactStatus,
+		},
+	};
+}
+
+export function buildLifecycleActions(graph: AgentPanelCanonicalSource): AgentPanelActionDescriptor[] {
+	const actions: AgentPanelActionDescriptor[] = [];
+
+	if (graph.lifecycle.actionability.canResume) {
+		actions.push({
+			id: AGENT_PANEL_ACTION_IDS.status.resume,
+			label: "Resume",
+			state: "enabled",
+		});
+	}
+
+	if (graph.lifecycle.actionability.canRetry) {
+		actions.push({
+			id: AGENT_PANEL_ACTION_IDS.status.retry,
+			label: "Retry",
+			state: "enabled",
+		});
+	}
+
+	if (graph.lifecycle.actionability.canArchive) {
+		actions.push({
+			id: AGENT_PANEL_ACTION_IDS.status.archive,
+			label: "Archive",
+			state: "enabled",
+		});
+	}
+
+	return actions;
+}
+
+export function displaySafeDegradationReason(
+	reason: OperationDegradationReason | null | undefined
+): string {
+	if (reason === null || reason === undefined) {
+		return "Tool operation is degraded.";
+	}
+
+	if (reason.code === "classification_failure") {
+		return "Tool operation could not be classified safely.";
+	}
+	if (reason.code === "missing_evidence") {
+		return "Tool operation is missing canonical evidence.";
+	}
+	if (reason.code === "absent_from_history") {
+		return "Tool operation is absent from provider history.";
+	}
+	if (reason.code === "invalid_provenance_key") {
+		return "Tool operation has invalid provenance.";
+	}
+	return "Tool operation has an impossible state transition.";
+}
